+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.2.4 (Unreleased)
 ### Added
 * Add diff to operation output.
+* Add replace limit option.
 
 # v0.2.3 (2020-07-18)
 ### Changed